Create comment reworked
